src/main/java/com/example/medicalmanagement/service/CustomUserDetailsService.java
package com.example.medicalmanagement.service;

import com.example.medicalmanagement.exceptionhandlers.NotFoundException;
import com.example.medicalmanagement.model.User;
import com.example.medicalmanagement.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.stereotype.Service;

@Service
public class CustomUserDetailsService implements UserDetailsService {
    @Autowired
    private UserRepository userRepository;


    public UserDetails loadUserByUsername(String username) throws NotFoundException {
       User user=userRepository.findByEmail(username);
        if (user == null) {
            throw new NotFoundException("User not found with email: " + username);
        }

        return org.springframework.security.core.userdetails.User.builder().username(user.getEmail()).password(user.getPassword()).roles(user.getRole()).build();
    }
}
